Codigo_Update_Monthly_Aena/requirements.txt
asn1crypto==1.5.1
async-generator==1.10
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
gspread==5.7.2
gspread-dataframe==3.3.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.1
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pipreqs==0.4.11
protobuf==4.21.12
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pygsheets==2.0.6
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.0.0
snowflake-sqlalchemy==1.4.6
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
webdriver-manager==3.8.5
wincertstore==0.2
wsproto==1.2.0
yarg==0.1.9
